requirements.txt
streamlit>=1.30.0
tensorflow>=2.10.0
numpy>=1.23.0
opencv-python>=4.8.0
pillow>=10.0.0